+++ SucculentAndPlantApp/Support/Extensions/Date+StripTime.swift
@@ -1,0 +1,16 @@
+//
+//  Date+StripTime.swift
+//  SucculentAndPlantApp
+//
+//  Created by Alek Michelson on 12/19/23.
+//
+
+import Foundation
+
+extension Date {
+    func stripTime() -> Date {
+        let components = Calendar.current.dateComponents([.year, .month, .day], from: self)
+        let date = Calendar.current.date(from: components)
+        return date!
+    }
+}
