+++ SucculentAndPlantApp/Support/Extensions/FIleManager+Image.swift
@@ -26,7 +26,7 @@ extension FileManager {
                 let url = URL.documentsDirectory.appendingPathComponent("\(id).jpg")
                 try data.write(to: url)
             } catch {
-                print(error.localizedDescription)
+                Logger.plantPal.error("\(#function) \(error.localizedDescription)")
             }
         } else {
             Logger.plantPal.error("Could not save image: \(id)")
@@ -42,7 +42,7 @@ extension FileManager {
                 print(error.localizedDescription)
             }
         } else {
-            Logger.plantPal.error("Cannot delete imaage \(id): image does not exist")
+            Logger.plantPal.error("\(#function) Could not delete image: \(id)")
         }
     }
 }
